Seperating Lambda API into core package and clid
