+++ workspaces/templates-lib/packages/template-lambda-api-cli/README.md
@@ -1,0 +1,3 @@
+# Goldstack Lambda API Template Tools
+
+This library contains an API and CLI for building Node.js modules that are deployed as AWS Lambda and expose an API on AWS.
