+++ workspaces/templates-lib/packages/template-lambda-api-cli/bin/template
@@ -4,7 +4,7 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 require('source-map-support').install();
 
-require('./../dist/src/templateLambdaApi')
+require('./../dist/src/templateLambdaApiCli')
   .run(process.argv)
   .catch((e) => {
     console.log(e);
