+++ workspaces/templates-lib/packages/template-lambda-api-cli/bin/template.ts
@@ -1,4 +1,4 @@
-import { run } from './../dist/src/templateLambdaApi';
+import { run } from './../dist/src/templateLambdaApiCli';
 
 run(process.argv).catch((e) => {
   console.log(e);
