+++ workspaces/templates-lib/packages/template-lambda-api-cli/jest.config.js
@@ -1,0 +1,12 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const base = require('../../../templates/jest.config');
+
+module.exports = {
+  ...base,
+  resolver: require.resolve('jest-pnp-resolver'),
+  globals: {
+    'ts-jest': {
+      tsconfig: 'tsconfig.json',
+    },
+  },
+};
