+++ workspaces/templates-lib/packages/template-lambda-api-cli/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@goldstack/template-lambda-api",
-  "version": "0.4.41",
+  "name": "@goldstack/template-lambda-api-cli",
+  "version": "0.4.0",
   "description": "Companion for project templates for deploying Lambdas as a REST API",
   "keywords": [
     "goldstack",
@@ -23,7 +23,11 @@
   "license": "MIT",
   "author": "Max Rohde",
   "sideEffects": false,
-  "main": "src/templateLambdaApi.ts",
+  "main": "src/templateLambdaApiCli.ts",
+  "bin": {
+    "template": "./bin/template",
+    "template-lambda-api": "./bin/template"
+  },
   "scripts": {
     "build": "yarn clean && yarn compile",
     "build:watch": "yarn clean && yarn compile-watch",
@@ -44,9 +48,24 @@
   "dependencies": {
     "@goldstack/infra": "0.3.36",
     "@goldstack/infra-aws": "0.3.44",
+    "@goldstack/template-lambda-api": "0.4.41",
+    "@goldstack/utils-aws-lambda": "0.1.28",
+    "@goldstack/utils-cli": "0.2.13",
+    "@goldstack/utils-config": "0.3.33",
+    "@goldstack/utils-docker": "0.3.38",
+    "@goldstack/utils-log": "0.2.14",
     "@goldstack/utils-package": "0.3.38",
+    "@goldstack/utils-package-config": "0.3.38",
+    "@goldstack/utils-s3-deployment": "0.3.55",
+    "@goldstack/utils-sh": "0.4.31",
+    "@goldstack/utils-template": "0.3.35",
     "@goldstack/utils-terraform": "0.3.68",
-    "source-map-support": "^0.5.21"
+    "@goldstack/utils-terraform-aws": "0.3.67",
+    "@yarnpkg/esbuild-plugin-pnp": "^3.0.0-rc.1",
+    "archiver": "^5.3.1",
+    "esbuild": "^0.14.34",
+    "source-map-support": "^0.5.21",
+    "yargs": "^15.4.1"
   },
   "devDependencies": {
     "@goldstack/utils-docs-cli": "0.3.11",
@@ -65,6 +84,6 @@
     "typescript": "^4.6.4"
   },
   "publishConfig": {
-    "main": "dist/src/templateLambdaApi.js"
+    "main": "dist/src/templateLambdaApiCli.js"
   }
 }
