+++ workspaces/templates-lib/packages/template-lambda-api-cli/scripts/buildTestLambdas.ts
