+++ workspaces/templates-lib/packages/template-lambda-api-cli/testData/build-test/customBuild.ts
