+++ workspaces/templates/packages/serverless-api/package.json
@@ -30,7 +30,7 @@
     "source-map-support": "^0.5.21"
   },
   "devDependencies": {
-    "@goldstack/template-lambda-api": "0.4.41",
+    "@goldstack/template-lambda-api-cli": "0.4.0",
     "@goldstack/utils-aws-http-api-local": "0.1.30",
     "@jest-mock/express": "^1.4.5",
     "@types/aws-lambda": "^8.10.88",
