+++ workspaces/templates/packages/serverless-api/scripts/template.ts
@@ -1,4 +1,4 @@
-import { run } from '@goldstack/template-lambda-api';
+import { run } from '@goldstack/template-lambda-api-cli';
 
 run(process.argv).catch((e) => {
   console.log(e);
